refactor(proxy): move pure error-classification helpers out of service.py
service.py had drifted 4 lines past the proxy-architecture line-count
gate (2604 vs. the 2600 limit), breaking CI for any PR regardless of
its actual diff. Move four self-contained, dependency-light helper
functions (_is_account_neutral_error_code, _is_local_account_cap_code,
_http_error_status_from_payload,
_openai_error_envelope_from_response_failed_payload) to
_service/support.py, which service.py is already allowed to import
from, and re-export them so every existing _facade()-based caller is
unaffected. Pure mechanical move, no behavior change.
